make the tab-switching function more generic and move it to utils.js
This will enable it's reuse in the image-detail section.
